created new Calculator class refactored Main class:  - moved functions to new Calculator class  - changed else if blocks into pattern matching
